fix MediaFile caching

+++ lib/src/main/java/scal/io/liger/model/MediaFile.java
@@ -53,33 +53,42 @@ public class MediaFile {
         return thumbnailFilePath;
     }
 
-    public void setThumbnailFilePath(String thumbnailFilePath) {
-        this.thumbnailFilePath = thumbnailFilePath;
-    }
-
-    public Bitmap getThumbnail() { // todo: disk cache, multiple sizes
+    public Bitmap getThumbnail() { // TODO: disk cache, multiple sizes
         Bitmap thumbnail = null;
 
         if (thumbnailFilePath == null) {
             if (medium.equals(Constants.VIDEO)) {
                 try {
-                    thumbnail = ThumbnailUtils.createVideoThumbnail(path, MediaStore.Images.Thumbnails.MINI_KIND);
                     //Log.d(" *** TESTING *** ", "CREATING NEW THUMBNAIL FILE FOR " + path);
 
-                    String thumbnailString = path.substring(0, path.lastIndexOf(File.separator) + 1) + UUID.randomUUID().toString() + ".png";
-                    File thumbnailFile = new File(thumbnailString);
-                    FileOutputStream thumbnailStream = new FileOutputStream(thumbnailFile);
+                    // FIXME should not be stored in the source location, but a cache dir in our app folder on the sd or internal cache if there is no SD
+                    // FIXME need to check datestamp on original file to check if our thumbnail is up to date
+                    // FIXME this should be run from a background thread as it does disk access  
+                    File originalFile = new File(path);
+                    String fileName = originalFile.getName();
+                    String[] tokens = fileName.split("\\.(?=[^\\.]+$)");
+                    thumbnailFilePath = path.substring(0, path.lastIndexOf(File.separator) + 1) + tokens[0] + "_thumbnail.png";
+                    File thumbnailFile = new File(thumbnailFilePath);
+                    if (thumbnailFile.exists()) {
+                        thumbnail = BitmapFactory.decodeFile(thumbnailFilePath);
+                    } else {
+                        FileOutputStream thumbnailStream = new FileOutputStream(thumbnailFile);
 
-                    thumbnail.compress(Bitmap.CompressFormat.PNG, 75, thumbnailStream);
-                    thumbnailStream.flush();
-                    thumbnailStream.close();
+                        thumbnail = ThumbnailUtils.createVideoThumbnail(path, MediaStore.Images.Thumbnails.MINI_KIND);
+                        thumbnail.compress(Bitmap.CompressFormat.PNG, 75, thumbnailStream); // FIXME make compression level configurable
+                        thumbnailStream.flush();
+                        thumbnailStream.close();
+                    }
 
-                    thumbnailFilePath = thumbnailString;
                     //Log.d(" *** TESTING *** ", "THUMBNAIL FILE SAVED AS " + thumbnailFilePath);
                 } catch (IOException ioe) {
                     //Log.d(" *** TESTING *** ", "EXCEPTION: " + ioe.getMessage());
                     return null;
                 }
+            } else if (medium.equals(Constants.AUDIO)) {
+                // TODO create audio thumbnails
+            } else if (medium.equals(Constants.PHOTO)) {
+                // TODO return the image directly
             } else {
                 Log.e(this.getClass().getName(), "can't create thumbnail file for " + path + ", unsupported medium: " + medium);
             }
